Exception advice, filter 추가 및 API 적용

+++ src/main/java/com/example/meetup_study/admin/AdminController.java
@@ -1,17 +1,22 @@
 package com.example.meetup_study.admin;
 
 import com.example.meetup_study.Category.CategoryService;
+import com.example.meetup_study.auth.exception.AccessTokenInvalidRequestException;
 import com.example.meetup_study.auth.jwt.JwtService;
 import com.example.meetup_study.hostReview.HostReviewService;
 import com.example.meetup_study.hostReview.domain.HostReview;
+import com.example.meetup_study.hostReview.exception.HostReviewInvalidRequestException;
+import com.example.meetup_study.hostReview.exception.HostReviewNotFoundException;
 import com.example.meetup_study.review.ReviewService;
 import com.example.meetup_study.review.domain.Review;
 import com.example.meetup_study.room.RoomService;
 import com.example.meetup_study.room.domain.Room;
 import com.example.meetup_study.room.domain.dto.RequestDeleteRoomDto;
 import com.example.meetup_study.room.domain.dto.RoomDto;
+import com.example.meetup_study.room.exception.RoomNotFoundException;
 import com.example.meetup_study.user.UserService;
 import com.example.meetup_study.user.domain.User;
+import com.example.meetup_study.user.fakeUser.exception.UserNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -39,15 +44,18 @@ public class AdminController {
         Optional<Long> userId = jwtService.extractUserId(accessToken);
 
         if(!userId.isPresent()){
-            throw new IllegalArgumentException("토큰에 유저가 없습니다.");
+            throw new AccessTokenInvalidRequestException();
         }
 
         Optional<User> userOpt = userService.findById(userId.get());
+        if(!userOpt.isPresent()){
+            throw new UserNotFoundException();
+        }
 
         Optional<Room> roomOpt = roomService.getRoom(requestDeleteRoomDto.getId());
 
-        if(!userOpt.isPresent() || !roomOpt.isPresent()){
-            throw new IllegalArgumentException("이 유저는 없거나, 방이 없거나, 방을 만들지 않았습니다.");
+        if(!roomOpt.isPresent()){
+            throw new RoomNotFoundException();
         }
 
         Optional<RoomDto> deletedRoomDto = adminService.deleteRoom(requestDeleteRoomDto.getId(), userOpt.get().getId());
@@ -63,13 +71,13 @@ public class AdminController {
         Optional<Long> userIdOpt = jwtService.extractUserId(accessToken);
 
         if(!userIdOpt.isPresent()){
-            throw new IllegalArgumentException("유저가 없습니다.");
+            throw new AccessTokenInvalidRequestException();
         }
 
         Optional<User> userOpt = userService.findById(userIdOpt.get());
 
         if(!userOpt.isPresent()){
-            throw new IllegalArgumentException("이 유저는 없거나, 방에 참여하지 않았습니다.");
+            throw new UserNotFoundException();
         }
 
         Optional<Review> reviewOpt = reviewService.findById(reviewId);
@@ -77,8 +85,8 @@ public class AdminController {
         Long roomId = reviewOpt.get().getRoom().getId();
 
         Optional<Room> roomOpt = roomService.getRoom(roomId);
-        if (!roomOpt.isPresent()) {
-            throw new IllegalArgumentException("방이 없습니다.");
+        if(!roomOpt.isPresent()){
+            throw new RoomNotFoundException();
         }
 
 
@@ -95,15 +103,19 @@ public class AdminController {
         Optional<Long> userIdOpt = jwtService.extractUserId(accessToken);
 
         if(!userIdOpt.isPresent()){
-            throw new IllegalArgumentException("유저가 없습니다.");
+            throw new AccessTokenInvalidRequestException();
         }
 
         Optional<User> userOpt = userService.findById(userIdOpt.get());
 
+        if(!userOpt.isPresent()){
+            throw new UserNotFoundException();
+        }
+
         Optional<HostReview> hostReviewOpt = adminService.deleteHostReview(hostReviewId, userOpt.get().getId());
 
         if(!hostReviewOpt.isPresent()){
-            throw new IllegalArgumentException("리뷰를 삭제하지 못했습니다.");
+            throw new HostReviewNotFoundException();
         }
 
         return ResponseEntity.ok(hostReviewOpt.get());
